# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop GazTochka API",
        "description": "Документация API для интернет-магазина ГазТочка",
        "contact": {
            "email": "admin@gaztochka.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить список всех товаров",
                "operationId": "e422a6932f9da90d47891a91e88effc1",
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создать новый товар",
                "operationId": "9bcbb575942cd412f2ea4306860f9636",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Товар успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить информацию о товаре по ID",
                "operationId": "bf64c626ced82f402f00356b4078b07c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о товаре",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновить существующий товар",
                "operationId": "9e35c066f44b11451b7bc4e3e7d47e8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Товар успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Товар не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удалить товар",
                "operationId": "862b06a891723d01d8eef100927f2e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Товар успешно удален"
                    },
                    "404": {
                        "description": "Товар не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductDTO": {
                "required": [
                    "category_id",
                    "name",
                    "slug"
                ],
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Газовый котел"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gazoviy-kotel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание товара",
                        "nullable": true
                    },
                    "attributes": {
                        "type": "object",
                        "example": {
                            "power": "24kW"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Газовый котел"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gazoviy-kotel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание товара",
                        "nullable": true
                    },
                    "attributes": {
                        "type": "object",
                        "example": {
                            "power": "24kW"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}